metering, audit: the counts on two result structs stop being int (#621)
* metering, audit: the counts on two result structs stop being int

Five exported result counts spelled themselves int while every other one
in the module — Deleted, Removed, Scanned, Leased, Pending, Ready, Count
and every Sweep and Reap return — answers in int64. A count's type is
exported API in the one way its value never shows: nothing a caller
writes against Flushed or Checked reads differently at either width, so
the narrowing is invisible right up until the day the field has to widen,
and widening an exported field after the tag is a /v15.

FlushResult.Claimed goes with Flushed, Skipped and Failed. The issue
named three of the four, but Claimed is the same field with the same
lifetime — it is assigned from a len, subtracted from the other two in
the backlog gauge, and sat beside Quantity and EventsReaped, which were
int64 already. Leaving it would have left the one struct spelling the
same fact two ways and the gauge converting one operand of a subtraction.

The conversions the widening removes are the evidence it was owed:
audit's gRPC converters were casting Checked in both directions because
the proto field always was an int64, and the backlog gauge was wrapping
its subtraction to reach a Record that always wanted one.

Each package pins the widths with a reflection test over its result
types, so a count added later at int fails here rather than in a
consumer's next major bump.

Closes #585


* countwidth: the rule gets one home, and the result types come to it

The width check landed as two copies, one in metering's test file and one in
audit's, each naming the same nine disallowed reflect kinds beside its own
result type. A list spelled twice is a list that can drift — the day one copy
learns about a kind the other does not, nothing says so and the older one goes
on passing — and that is the shape this module extracts: not what is merely
written twice, but what can be wrong twice.

internal/countwidth names the kinds once and rosters the types that come to
them: metering.FlushResult, audit.VerificationResult and audit.Break. Both
per-package copies are gone.

The roster is also what turns "every exported count" into a claim. A result type
added to a package with no convention test of its own was a type nothing
checked, and stayed that way until somebody remembered; now it is a missing row
in the file where the rows that are already right live.

The assertion stays over every exported field rather than over a list of the
ones that count, and the message says why rather than asserting the field counts:
a field this test has never seen is exactly the one a widening elsewhere would
have missed, and a roster of field names would need extending by whoever added
the narrow field.

Confirmed to red by adding an exported int to a rostered type, which it reports
by package, type, field and kind. Narrowing an existing count could not be used
to check it — metering does not compile with Claimed at int, which is its own
small evidence the widening was owed.

internal/tiercheck gets the roster row the new package owes it.


---------

+++ metering/flusher.go
@@ -49,18 +49,18 @@ var ErrFlusherPanicked = platformerrors.New("metering flush panicked")
 // FlushResult is what one pass did.
 type FlushResult struct {
 	// Claimed is how many totals the pass leased.
-	Claimed int
+	Claimed int64
 
 	// Flushed is how many were posted to the provider and settled.
-	Flushed int
+	Flushed int64
 
 	// Skipped is how many had no provider ref and were settled without a post
 	// — a subject on a plan that does not bill for that meter. Not a failure; see
 	// ErrNoProviderRef.
-	Skipped int
+	Skipped int64
 
 	// Failed is how many could not be posted and were returned for retry.
-	Failed int
+	Failed int64
 
 	// Quantity is the total usage posted, summed across every successful flush.
 	Quantity int64
@@ -256,7 +256,7 @@ func (f *Flusher) Flush(ctx context.Context) (*FlushResult, error) {
 	if err != nil {
 		errs = append(errs, platformerrors.Wrap(err, "claiming flushable metering totals"))
 	} else {
-		result.Claimed = len(claimed)
+		result.Claimed = int64(len(claimed))
 		f.post(ctx, claimed, result)
 	}
 
@@ -265,7 +265,7 @@ func (f *Flusher) Flush(ctx context.Context) (*FlushResult, error) {
 	// revenue that is not being invoiced, and no counter of successful flushes can
 	// distinguish "flushing steadily" from "flushing steadily while a queue builds
 	// behind it".
-	f.backlogGauge.Record(ctx, int64(result.Claimed-result.Flushed-result.Skipped))
+	f.backlogGauge.Record(ctx, result.Claimed-result.Flushed-result.Skipped)
 
 	if !f.cfg.DisableReap {
 		reaped, reapErr := f.store.ReapEvents(ctx, now.Add(-f.cfg.EventRetention), f.cfg.ReapBatchSize)
